Package Generation

Package Generation

Creating new packages with PkgTemplates revolves around creating a new Template, then calling generate on it.

Template

PkgTemplates.Template — Type.

Template(; kwargs...) -> Template

Records common information used to generate a package. If you don't wish to manually create a template, you can use interactive_template instead.

Keyword Arguments

user::AbstractString=""): GitHub username. If left unset, it will try to take the value of a supplied git config's "github.user" key, then the global git config's value. If neither is set, an ArgumentError is thrown. This is case-sensitive for some plugins, so take care to enter it correctly.

host::AbstractString="github.com": URL to the code hosting service where your package will reside. Note that while hosts other than GitHub won't cause errors, they are not officially supported and they will cause certain plugins will produce incorrect output. For example, AppVeyor's badge image will point to a GitHub-specific URL, regardless of the value of host.

license::Union{AbstractString, Void}="MIT": Name of the package license. If no license is specified, no license is created. show_license can be used to list all available licenses, or to print out a particular license's text.

authors::Union{AbstractString, Array}="": Names that appear on the license. Supply a string for one author, and an array for multiple. Similarly to user, it will try to take the value of a supplied git config's "user.name" key, then the global git config's value, if it is left unset

years::Union{Int, AbstractString}=Int(Dates.year(now())): Copyright years on the license. Can be supplied by a number, or a string such as "2016 - 2017".

dir::AbstractString=Pkg.dir(): Directory in which the package will go.

julia_version::VersionNumber=VERSION: Minimum allowed Julia version.

requirements::Vector{String}=String[]: Package requirements. If there are duplicate requirements with different versions, i.e. ["PkgTemplates", "PkgTemplates 0.1"], an ArgumentError is thrown. Each entry in this array will be copied into the REQUIRE file of packages generated with this template.

git_config::Dict{String, String}=Dict{String, String}(): Git configuration options.

plugins::Plugin[]: A list of Plugins that the package will include.

Notes

When you create a Template, a temporary directory is created with mktempdir(). This directory will be removed after you call generate. Creating multiple packages in succession with the same instance of a template will still work, but there is a miniscule chance of another process sharing the temporary directory, which could result in the created package repository containing untracked files that don't belong.

source

PkgTemplates.interactive_template — Function.

interactive_template(; fast::Bool) -> Template

Interactively create a Template. If fast is set, defaults will be assumed for all values except username and plugins.

source

generate

PkgTemplates.generate — Function.

generate(
    pkg_name::AbstractString,
    t::Template;
    force::Bool=false,
    ssh::Bool=false,
) -> Void

Generate a package names pkg_name from template.

Keyword Arguments

force::Bool=false: Whether or not to overwrite old packages with the same name.

ssh::Bool=false: Whether or not to use SSH for the remote.

Notes

The package is generated entirely in a temporary directory (t.temp_dir), and only moved into joinpath(t.dir, pkg_name) at the very end. In the case of an error, the temporary directory will contain leftovers, but the destination directory will remain untouched (this is especially helpful when force=true).

source

Helper Functions

gen_entrypoint

PkgTemplates.gen_entrypoint — Function.

gen_entrypoint(pkg_name::AbstractString, template::Template) -> Vector{String}

Create the module entrypoint in the temp package directory.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose entrypoint we are generating.

Returns an array of generated file/directory names.

source

gen_tests

PkgTemplates.gen_tests — Function.

gen_tests(pkg_name::AbstractString, template::Template) -> Vector{String}

Create the test directory and entrypoint in the temp package directory.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose tests we are generating.

Returns an array of generated file/directory names.

source

gen_require

PkgTemplates.gen_require — Function.

gen_require(pkg_name::AbstractString, template::Template) -> Vector{String}

Create the REQUIRE file in the temp package directory.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose REQUIRE we are generating.

Returns an array of generated file/directory names.

source

gen_readme

PkgTemplates.gen_readme — Function.

gen_readme(pkg_name::AbstractString, template::Template) -> Vector{String}

Create a README in the temp package directory with badges for each enabled plugin.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose README we are generating.

Returns an array of generated file/directory names.

source

gen_gitignore

PkgTemplates.gen_gitignore — Function.

gen_gitignore(pkg_name::AbstractString, template::Template) -> Vector{String}

Create a .gitignore in the temp package directory.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose .gitignore we are generating.

Returns an array of generated file/directory names.

source

gen_license

PkgTemplates.gen_license — Function.

gen_license(pkg_name::AbstractString, template::Template) -> Vector{String}

Create a license in the temp package directory.

Arguments

pkg_name::AbstractString: Name of the package.

template::Template: The template whose LICENSE we are generating.

Returns an array of generated file/directory names.

source