Plugins

Plugins

Plugins are the driver for PkgTemplates's customization and extension. This page describes plugins that already exist; for information on writing your own plugins, see Plugin Development.

TravisCI

PkgTemplates.TravisCI — Type.

TravisCI(; config_file::Union{AbstractString, Void}="") -> TravisCI

Add TravisCI to a template's plugins to add a .travis.yml configuration file to generated repositories, and an appropriate badge to the README.

Keyword Arguments:

config_file::Union{AbstractString, Void}="": Path to a custom .travis.yml. If nothing is supplied, no file will be generated.

source

AppVeyor

PkgTemplates.AppVeyor — Type.

AppVeyor(; config_file::Union{AbstractString, Void}="") -> AppVeyor

Add AppVeyor to a template's plugins to add a .appveyor.yml configuration file to generated repositories, and an appropriate badge to the README.

Keyword Arguments

config_file::Union{AbstractString, Void}="": Path to a custom .appveyor.yml. If nothing is supplied, no file will be generated.

source

CodeCov

PkgTemplates.CodeCov — Type.

CodeCov(; config_file::Union{AbstractString, Void}="") -> CodeCov

Add CodeCov to a template's plugins to add a .codecov.yml configuration file to generated repositories, and an appropriate badge to the README. Also updates the .gitignore accordingly.

Keyword Arguments:

config_file::Union{AbstractString, Void}="": Path to a custom .codecov.yml. If nothing is supplied, no file will be generated.

source

Coveralls

PkgTemplates.Coveralls — Type.

Coveralls(; config_file::Union{AbstractString, Void}="") -> Coveralls

Add Coveralls to a template's plugins to optionally add a .coveralls.yml configuration file to generated repositories, and an appropriate badge to the README. Also updates the .gitignore accordingly.

Keyword Arguments:

config_file::Union{AbstractString, Void}=nothing: Path to a custom .coveralls.yml. If left unset, no file will be generated.

source

Documenter

PkgTemplates.Documenter — Type.

Add a Documenter subtype to a template's plugins to add support for documentation generation via Documenter.jl.

source

GitHubPages

PkgTemplates.GitHubPages — Type.

GitHubPages(; assets::Vector{<:AbstractString}=String[]) -> GitHubPages

Add GitHubPages to a template's plugins to add Documenter support via GitHub Pages, including automatic uploading of documentation from TravisCI. Also adds appropriate badges to the README, and updates the .gitignore accordingly.

Keyword Arguments

assets::Vector{String}=String[]: Array of paths to Documenter asset files.

source