User Guide

PkgTemplates User Guide

PkgTemplates User Guide

Template

Plugins

A More Complicated Example

Custom Template Files

Extending Existing Plugins

Saving Templates

Using PkgTemplates is straightforward. Just create a Template, and call it on a package name to generate that package:

using PkgTemplates
t = Template()
t("MyPkg")

Template

PkgTemplates.Template — Type.

Template(; kwargs...)

A configuration used to generate packages.

Keyword Arguments

User Options

user::AbstractString="username": GitHub (or other code hosting service) username. The default value comes from the global Git config (github.user). If no value is obtained, many plugins that use this value will not work.

authors::Union{AbstractString, Vector{<:AbstractString}}="name <email>": Package authors. Like user, it takes its default value from the global Git config (user.name and user.email).

Package Options

dir::AbstractString="~/.julia/dev": Directory to place packages in.

host::AbstractString="github.com": URL to the code hosting service where packages will reside.

julia::VersionNumber=v"1.0.0": Minimum allowed Julia version.

Template Plugins

plugins::Vector{<:Plugin}=Plugin[]: A list of Plugins used by the template.

disable_defaults::Vector{DataType}=DataType[]: Default plugins to disable. The default plugins are ProjectFile, SrcDir, Tests, Readme, License, and Git. To override a default plugin instead of disabling it altogether, supply it via plugins.

To create a package from a Template, use the following syntax:

julia> t = Template();

julia> t("PkgName")

source

Plugins

Plugins add functionality to Templates. There are a number of plugins available to automate common boilerplate tasks.

Default Plugins

These plugins are included by default. They can be overridden by supplying another value via the plugins keyword, or disabled by supplying the type via the disable_defaults keyword.

PkgTemplates.ProjectFile — Type.

ProjectFile(; version=v"0.1.0")

Creates a Project.toml.

Keyword Arguments

version::VersionNumber: The initial version of created packages.

source

PkgTemplates.SrcDir — Type.

SrcDir(; file="~/build/invenia/PkgTemplates.jl/templates/src/module.jl")

Creates a module entrypoint.

Keyword Arguments

file::AbstractString: Template file for src/<module>.jl.

source

PkgTemplates.Tests — Type.

Tests(; file="~/build/invenia/PkgTemplates.jl/templates/test/runtests.jl", project=false)

Sets up testing for packages.

Keyword Arguments

file::AbstractString: Template file for runtests.jl.

project::Bool: Whether or not to create a new project for tests (test/Project.toml). See here for more details.

Note

Managing test dependencies with test/Project.toml is only supported in Julia 1.2 and later.

source

PkgTemplates.Readme — Type.

Readme(;
    file="~/build/invenia/PkgTemplates.jl/templates/README.md",
    destination="README.md",
    inline_badges=false,
)

Creates a README file that contains badges for other included plugins.

Keyword Arguments

file::AbstractString: Template file for the README.

destination::AbstractString: File destination, relative to the repository root. For example, values of "README" or "README.rst" might be desired.

inline_badges::Bool: Whether or not to put the badges on the same line as the package name.

source

PkgTemplates.License — Type.

License(; name="MIT", path=nothing, destination="LICENSE")

Creates a license file.

Keyword Arguments

name::AbstractString: Name of a license supported by PkgTemplates. Available licenses can be seen here.

path::Union{AbstractString, Nothing}: Path to a custom license file. This keyword takes priority over name.

destination::AbstractString: File destination, relative to the repository root. For example, "LICENSE.md" might be desired.

source

PkgTemplates.Git — Type.

Git(; ignore=String[], ssh=false, manifest=false, gpgsign=false)

Creates a Git repository and a .gitignore file.

Keyword Arguments

ignore::Vector{<:AbstractString}: Patterns to add to the .gitignore. See also: gitignore.

ssh::Bool: Whether or not to use SSH for the remote. If left unset, HTTPS is used.

manifest::Bool: Whether or not to commit Manifest.toml.

gpgsign::Bool: Whether or not to sign commits with your GPG key. This option requires that the Git CLI is installed, and for you to have a GPG key associated with your committer identity.

source

Continuous Integration (CI)

These plugins will create the configuration files of common CI services for you.

PkgTemplates.AppVeyor — Type.

AppVeyor(;
    file="~/build/invenia/PkgTemplates.jl/templates/appveyor.yml",
    x86=false,
    coverage=true,
    extra_versions=["1.0", "1.0", "nightly"],
)

Integrates your packages with AppVeyor via AppVeyor.jl.

Keyword Arguments

file::AbstractString: Template file for .appveyor.yml.

x86::Bool: Whether or not to run builds on 32-bit systems, in addition to the default 64-bit builds.

coverage::Bool: Whether or not to publish code coverage. Codecov must also be included.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

source

PkgTemplates.CirrusCI — Type.

CirrusCI(;
    file="~/build/invenia/PkgTemplates.jl/templates/cirrus.yml",
    image="freebsd-12-0-release-amd64",
    coverage=true,
    extra_versions=["1.0", "1.0", "nightly"],
)

Integrates your packages with Cirrus CI via CirrusCI.jl.

Keyword Arguments

file::AbstractString: Template file for .cirrus.yml.

image::AbstractString: The FreeBSD image to be used.

coverage::Bool: Whether or not to publish code coverage. Codecov must also be included.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

Note

Code coverage submission from Cirrus CI is not yet supported by Coverage.jl.

source

PkgTemplates.DroneCI — Type.

DroneCI(;
    file="~/build/invenia/PkgTemplates.jl/templates/drone.star",
    amd64=true,
    arm=false,
    arm64=false,
    extra_versions=["1.0", "1.0"],
)

Integrates your packages with Drone CI.

Keyword Arguments

file::AbstractString: Template file for .drone.star.

destination::AbstractString: File destination, relative to the repository root. For example, you might want to generate a .drone.yml instead of the default Starlark file.

amd64::Bool: Whether or not to run builds on AMD64.

arm::Bool: Whether or not to run builds on ARM (32-bit).

arm64::Bool: Whether or not to run builds on ARM64.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

Note

Nightly Julia is not supported.

source

PkgTemplates.GitHubActions — Type.

GitHubActions(;
    file="~/build/invenia/PkgTemplates.jl/templates/github/workflows/ci.yml",
    destination="ci.yml",
    linux=true,
    osx=true,
    windows=true,
    x64=true,
    x86=false,
    coverage=true,
    extra_versions=["1.0", "1.0"],
)

Integrates your packages with GitHub Actions.

Keyword Arguments

file::AbstractString: Template file for the workflow file.

destination::AbstractString: Destination of the worflow file, relative to .github/workflows.

linux::Bool: Whether or not to run builds on Linux.

osx::Bool: Whether or not to run builds on OSX (MacOS).

windows::Bool: Whether or not to run builds on Windows.

x64::Bool: Whether or not to run builds on 64-bit architecture.

x86::Bool: Whether or not to run builds on 32-bit architecture.

coverage::Bool: Whether or not to publish code coverage. Another code coverage plugin such as Codecov must also be included.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

Note

If using coverage plugins, don't forget to manually add your API tokens as secrets, as described here.

Note

Nightly Julia is not supported.

source

PkgTemplates.GitLabCI — Type.

GitLabCI(;
    file="~/build/invenia/PkgTemplates.jl/templates/gitlab-ci.yml",
    coverage=true,
    extra_versions=["1.0", "1.0"],
)

Integrates your packages with GitLab CI.

Keyword Arguments

file::AbstractString: Template file for .gitlab-ci.yml.

coverage::Bool: Whether or not to compute code coverage.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

GitLab Pages

Documentation can be generated by including a Documenter{GitLabCI} plugin. See Documenter for more information.

Note

Nightly Julia is not supported.

source

PkgTemplates.TravisCI — Type.

TravisCI(;
    file="~/build/invenia/PkgTemplates.jl/templates/travis.yml",
    linux=true,
    osx=true,
    windows=true,
    x64=true,
    x86=false,
    arm64=false,
    coverage=true,
    extra_versions=["1.0", "1.0", "nightly"],
)

Integrates your packages with Travis CI.

Keyword Arguments

file::AbstractString: Template file for .travis.yml.

linux::Bool: Whether or not to run builds on Linux.

osx::Bool: Whether or not to run builds on OSX (MacOS).

windows::Bool: Whether or not to run builds on Windows.

x64::Bool: Whether or not to run builds on 64-bit architecture.

x86::Bool: Whether or not to run builds on 32-bit architecture.

arm64::Bool: Whether or not to run builds on the ARM64 architecture.

coverage::Bool: Whether or not to publish code coverage. Another code coverage plugin such as Codecov must also be included.

extra_versions::Vector: Extra Julia versions to test, as strings or VersionNumbers.

source

Code Coverage

These plugins will enable code coverage reporting from CI.

PkgTemplates.Codecov — Type.

Codecov(; file=nothing)

Sets up code coverage submission from CI to Codecov.

Keyword Arguments

file::Union{AbstractString, Nothing}: Template file for .codecov.yml, or nothing to create no file.

source

PkgTemplates.Coveralls — Type.

Coveralls(; file=nothing)

Sets up code coverage submission from CI to Coveralls.

Keyword Arguments

file::Union{AbstractString, Nothing}: Template file for .coveralls.yml, or nothing to create no file.

source

Documentation

PkgTemplates.Documenter — Type.

Documenter{T<:Union{TravisCI, GitLabCI, Nothing}}(;
    make_jl="~/build/invenia/PkgTemplates.jl/templates/docs/make.jl",
    index_md="~/build/invenia/PkgTemplates.jl/templates/docs/src/index.md",
    assets=String[],
    canonical_url=make_canonical(T),
    makedocs_kwargs=Dict{Symbol, Any}(),
)

Sets up documentation generation via Documenter.jl. Documentation deployment depends on T, where T is some supported CI plugin, or Nothing to only support local documentation builds.

Supported Type Parameters

GitHubActions: Deploys documentation to GitHub Pages with the help of GitHubActions.

TravisCI: Deploys documentation to GitHub Pages with the help of TravisCI.

GitLabCI: Deploys documentation to GitLab Pages with the help of GitLabCI.

Nothing (default): Does not set up documentation deployment.

Keyword Arguments

make_jl::AbstractString: Template file for make.jl.

index_md::AbstractString: Template file for index.md.

assets::Vector{<:AbstractString}: Extra assets for the generated site.

canonical_url::Union{Function, Nothing}: A function to generate the site's canonical URL. The default value will compute GitHub Pages and GitLab Pages URLs for TravisCI and GitLabCI, respectively. If set to nothing, no canonical URL is set.

makedocs_kwargs::Dict{Symbol}: Extra keyword arguments to be inserted into makedocs.

Note

If deploying documentation with Travis CI, don't forget to complete the required configuration.

source

Miscellaneous

PkgTemplates.Develop — Type.

Develop()

Adds generated packages to the current environment by deving them. See the Pkg documentation here for more details.

source

PkgTemplates.Citation — Type.

Citation(; file="~/build/invenia/PkgTemplates.jl/templates/CITATION.bib", readme=false)

Creates a CITATION.bib file for citing package repositories.

Keyword Arguments

file::AbstractString: Template file for CITATION.bib.

readme::Bool: Whether or not to include a section about citing in the README.

source

A More Complicated Example

Here are a few example templates that use the options and plugins explained above.

This one includes plugins suitable for a project hosted on GitHub, and some other customizations:

Template(; 
    user="my-username",
    dir="~/code",
    authors="Acme Corp",
    julia=v"1.1",
    plugins=[
        License(; name="MPL"),
        Git(; manifest=true, ssh=true),
        GitHubActions(; x86=true),
        Codecov(),
        Documenter{GitHubActions}(),
        Develop(),
    ],
)

Here's one that works well for projects hosted on GitLab:

Template(;
    user="my-username",
    host="gitlab.com",
    plugins=[
        GitLabCI(),
        Documenter{GitLabCI}(),
    ],
)

Custom Template Files

Templates vs Templating

This documentation refers plenty to Templates, the package's main type, but it also refers to "template files" and "text templating", which are plaintext files with placeholders to be filled with data, and the technique of filling those placeholders with data, respectively.

These concepts should be familiar if you've used Jinja or Mustache (Mustache is the particular flavour used by PkgTemplates, via Mustache.jl). Please keep the difference between these two things in mind!

Many plugins support a file argument or similar, which sets the path to the template file to be used for generating files. Each plugin has a sensible default that should make sense for most people, but you might have a specialized workflow that requires a totally different template file.

If that's the case, a basic understanding of Mustache's syntax is required. Here's an example template file:

Hello, {{{name}}}.

{{#weather}}
It's {{{weather}}} outside. 
{{/weather}}
{{^weather}}
I don't know what the weather outside is.
{{/weather}}

{{#has_things}}
I have the following things:
{{/has_things}}
{{#things}}
- Here's a thing: {{{.}}}
{{/things}}

{{#people}}
- {{{name}}} is {{{mood}}}
{{/people}}

In the first section, name is a key, and its value replaces {{{name}}}.

In the second section, weather's value may or may not exist. If it does exist, then "It's weather outside" is printed. Otherwise, "I don't know what the weather outside is" is printed. Mustache uses a notion of "truthiness" similar to Python or JavaScript, where values of nothing, false, or empty collections are all considered to not exist.

In the third section, has_things' value is printed if it's truthy. Then, if the things list is truthy (i.e. not empty), its values are each printed on their own line. The reason that we have two separate keys is that {{#things}} iterates over the whole things list, even when there are no {{{.}}} placeholders, which would duplicate "I have the following things:" n times.

The fourth section iterates over the people list, but instead of using the {{{.}}} placeholder, we have name and mood, which are keys or fields of the list elements. Most types are supported here, including Dicts and structs. NamedTuples require you to use {{{:name}}} instead of the normal {{{name}}}, though.

You might notice that some curlies are in groups of two ({{key}}), and some are in groups of three ({{{key}}}). Whenever we want to subtitute in a value, using the triple curlies disables HTML escaping, which we rarely want for the types of files we're creating. If you do want escaping, just use the double curlies. And if you're using different delimiters, for example <<foo>>, use <<&foo>> to disable escaping.

Assuming the following view:

struct Person; name::String; mood::String; end
things = ["a", "b", "c"]
view = Dict(
    "name" => "Chris",
    "weather" => "sunny",
    "has_things" => !isempty(things),
    "things" => things,
    "people" => [Person("John", "happy"), Person("Jane", "sad")],
)

Our example template would produce this:

Hello, Chris.

It's sunny outside.

I have the following things:
- Here's a thing: a
- Here's a thing: b
- Here's a thing: c

- John is happy
- Jane is sad

Extending Existing Plugins

Most of the existing plugins generate a file from a template file. If you want to use custom template files, you may run into situations where the data passed into the templating engine is not sufficient. In this case, you can look into implementing user_view to supply whatever data is necessary for your use case.

PkgTemplates.user_view — Function.

user_view(::Plugin, ::Template, pkg::AbstractString) -> Dict{String, Any}

The same as view, but for use by package users for extension.

Values returned by this function will override those from view when the keys are the same.

source

For example, suppose you were using the Readme plugin with a custom template file that looked like this:

# {{PKG}}

Created on *{{TODAY}}*.

The view function supplies a value for PKG, but it does not supply a value for TODAY. Rather than override view, we can implement this function to get both the default values and whatever else we need to add.

user_view(::Readme, ::Template, ::AbstractString) = Dict("TODAY" => today())

Saving Templates

One of the main reasons for PkgTemplates' existence is for new packages to be consistent. This means using the same template more than once, so we want a way to save a template to be used later.

Here's my recommendation for loading a template whenever it's needed:

function template()
    @eval using PkgTemplates
    Template(; #= ... =#)
end

Add this to your startup.jl, and you can create your template from anywhere, without incurring any startup cost.

Another strategy is to write the string representation of the template to a Julia file:

const t = Template(; #= ... =#)
open("template.jl", "w") do io
    println(io, "using PkgTemplates")
    sprint(show, io, t)
end

Then the template is just an include away:

const t = include("template.jl")

The only disadvantage to this approach is that the saved template is much less human-readable than code you wrote yourself.

One more method of saving templates is to simply use the Serialization package in the standard library:

const t = Template(; #= ... =#)
using Serialization
open(io -> serialize(io, t), "template.bin", "w")

Then simply deserialize to load:

using Serialization
const t = open(deserialize, "template.bin")

This approach has the same disadvantage as the previous one, and the serialization format is not guaranteed to be stable across Julia versions.